requirements.txt
PySide2
Pillow
PyMuPDF
PyYAML
obsub
pytest
pytest-qt